Got the editing functionality to work as intended
