+++ app/javascript/components/App.js
@@ -4,6 +4,7 @@ import CareerMainPage from './CareerMainPage'
 import CreateJob from './pages/jobs/CreateJob'
 import ShowCurrentJob from './pages/jobs/ShowCurrentJob'
 import 'bootstrap/dist/css/bootstrap.min.css';
+import { useState } from 'react'
 
 import Header from "./pages/Header"
 import { BrowserRouter as Router, Link, Switch, Route } from "react-router-dom"
@@ -15,6 +16,32 @@ function App(props) {
       sign_out_route,
       current_user_id
     } = props
+
+    const [ errors, setErrors ] = useState(null)
+    const [ apiJobsData, setApiJobsData ] = useState([])
+
+    function getJob() {
+        return fetch('/jobs')
+            .then( resp => {
+                if (resp.status === 200) {
+                    return resp.json()
+                } else {
+                    return Promise.new(() => {
+                        resolve({error: 'there was an error'})
+                    })
+                }
+            })
+    }
+
+    function loadJobs(){
+        getJob()
+            .then(jobs => {
+                if(jobs.errors) {
+                    setErrors(jobs.errors)
+                }
+                setApiJobsData(jobs)
+            })
+    }
 
     return (
         <Router>
@@ -32,12 +59,12 @@ function App(props) {
                         <Switch>
                             <Route exact path="/careermainpage">
                                 <CareerMainPage
-                                    current_user_id={current_user_id}
+                                    current_user_id={current_user_id} getJob= {getJob}
                                  />
                             </Route>
 
                             <Route exact path='/jobs/:id' >
-                                <ShowCurrentJob />
+                                <ShowCurrentJob getJob={getJob} loadJobs = {loadJobs}/>
                             </Route>
